db testing email is unnique
Testing on our database to make sure the email of the user trying to register is unique. Adding the "message" variable to the "connexion.ejs" and styling it with "connexion.css". Also doing the backend to make sure the apssword is unique. Plus before bigining everythin, downloaded, the bcryptjs module to hash passwords and jsonwebtoken which is not yet used.